Refactor the heading into its own component
- Add option to remove the back button in the header through a prop.
- Add the ability to pass in the title as a prop to the compojnent.
