+++ client/src/pages/account/report.tsx
@@ -1,12 +1,10 @@
-import Image from "next/image";
-import { useRouter } from "next/router";
-import React, { FormEvent, useState } from "react";
+import React, { useState } from "react";
 
 import { Button } from "@/components/ui/button";
+import Header from "@/components/ui/header";
 import WriteIn, { WriteInFormData } from "@/components/ui/write-in";
 
 export default function Report() {
-  const router = useRouter();
   const [formReady, setFormReady] = useState(false);
   const [formData, setFormData] = useState<WriteInFormData>({
     subject: "",
@@ -29,26 +27,12 @@ export default function Report() {
     }
   };
 
-  const handleBackClick = () => {
-    router.back();
-  };
-
   return (
     <div
       id="report"
       className="flex min-h-screen flex-col bg-penni-background-input-light-mode"
     >
-      <div className="flex flex-grow-0 flex-row items-center justify-center bg-penni-background-light-mode px-4 pb-4 pt-6">
-        <Button onClick={handleBackClick} variant="link">
-          <Image
-            src="/icons/arrow_back.svg"
-            alt="Back"
-            width={24}
-            height={24}
-          />
-        </Button>
-        <h2 className="body-medium flex-grow text-center">Report a problem</h2>
-      </div>
+      <Header title="Report a problem" />
       <div className="h-[0.12rem] grow-0 bg-penni-background-input-light-mode pt-3"></div>
       <div className="flex flex-grow flex-col overflow-auto">
         <form onSubmit={handleSubmit} className="flex flex-grow flex-col">
